requirements.txt
Django==1.9
Pillow==3.2.0
PyJWT==1.4.0
argparse==1.4.0
bleach==1.4.2
defusedxml==0.4.1
django-allauth==0.25.2
django-appconf==1.0.2
django-bleach==0.3.0
django-braces==1.8.1
django-crispy-forms==1.6.0
django-imagekit==3.3
django-markdown-deux==1.0.5
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.3.3
gunicorn==19.4.5
html5lib==0.9999999
markdown2==2.3.1
oauthlib==1.0.3
pilkit==1.1.13
psycopg2==2.6.1
python-dateutil==2.5.3
python-social-auth==0.2.18
python3-openid==3.0.10
pytz==2016.3
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
